Include unofficial New Relic configurations
New relic has some unofficial configuration that are not
documented. In the future this configurations could be
promoted to an official configuration an then it would
deserve an specific entry on our template. For now we
just included a generic hash option to allow this
configuration to be setted.

As an example, this feature could be used to set
configurations like ignore_reclaimable, proposed by
New Relic team on this post:

https://discuss.newrelic.com/t/memavailable-vs-memfree/12760/6